+++ MonsterGame_WithEngine/MonsterEngine/MonsterPoint2D.h
@@ -2,26 +2,27 @@
 #define POINT2D_H
 
 #include <math.h>
-
+#include "MMath.h"
+#include "assert.h"
 class MonsterPoint2D
 {
 public:
-	MonsterPoint2D(double init_x, double init_y);
-	inline double Magnitude() const;
-	inline double DotProduct(const MonsterPoint2D& other) const;
+	MonsterPoint2D(float init_x, float init_y);
+	inline float Magnitude() const;
+	inline float DotProduct(const MonsterPoint2D& other) const;
 	void Normalize();
-	inline double x() const;
-	inline double y() const;
-	inline void x(const double val);
-	inline void y(const double val);
+	inline float x() const;
+	inline float y() const;
+	inline void x(const float val);
+	inline void y(const float val);
 
 	inline MonsterPoint2D operator+(const MonsterPoint2D& rhs) const;
 	inline MonsterPoint2D operator-(const MonsterPoint2D& rhs) const;
 
 
 private:
-	double coord_x_;
-	double coord_y_;
+	float coord_x_;
+	float coord_y_;
 };
 #include "MonsterPoint2D-inl.h"
 #endif
