+++ MonsterGame_WithEngine/MonsterEngine/MonsterTransform.cpp
@@ -2,7 +2,7 @@
 
 
 
-MonsterTransform::MonsterTransform(double new_x, double new_y) :
+MonsterTransform::MonsterTransform(float new_x, float new_y) :
 	position_(new_x, new_y)
 {
 }
@@ -12,7 +12,7 @@ MonsterTransform::~MonsterTransform()
 {
 }
 
-void MonsterTransform::Translate(double delta_x, double delta_y)
+void MonsterTransform::Translate(float delta_x, float delta_y)
 {
 	position_.x(position_.x() + delta_x);
 	position_.y(position_.y() + delta_y);
